+++ test/apps/app/src/window.ts
@@ -133,7 +133,8 @@ Object.assign(window, {
           expireEarlySeconds,
           autoRenew: true,
           autoRemove: false,
-          syncStorage: config?.tokenManager?.syncStorage
+          syncStorage: config?.tokenManager?.syncStorage,
+          broadcastChannelName: config.clientId + '_crossTabTest'
         };
         config.isTokenRenewPage = true;