+++ test/spec/TokenManager/browser.ts
@@ -75,7 +75,10 @@ describe('TokenManager (browser)', function() {
     client.tokenManager.off('added');
     client.tokenManager.off('removed');
     if (start) {
+      util.disableLeaderElection();
+      util.mockLeader();
       client.tokenManager.start();
+      client.serviceManager.start();
     }
     return client;
   }
@@ -87,9 +90,8 @@ describe('TokenManager (browser)', function() {
   });
   
   afterEach(() => {
-    if (client && client.tokenManager) {
-      client.tokenManager.stop();
-    }
+    client?.tokenManager?.stop();
+    client?.serviceManager?.stop();
   });
 
   describe('options', () => {